requirements.txt
Flask
requests
openai
pixivpy3
python-dotenv
cloudinary
pymongo
